Fix regressing hits test
